server/requirements.txt
flask~=3.1.1
flask-cors~=6.0.1
gevent~=25.5.1
loguru~=0.7.3
jsonschema~=4.24.0
PyYAML~=6.0.2
Nuitka~=2.7.12
